Moved BiDAF into a reading_comprehension/ module
