+++ allennlp/models/reading_comprehension/__init__.py
@@ -1,0 +1,8 @@
+"""
+Reading comprehension is loosely defined as follows: given a question and a passage of text that
+contains the answer, answer the question.
+
+These submodules contain models for things that are predominantly focused on reading comprehension.
+"""
+
+from allennlp.models.reading_comprehension.bidaf import BidirectionalAttentionFlow
